ConversationDataManagerの変更
ConversationDataManagerの変更
